fix: docker node BE build
