+++ Dockerfile.node
@@ -2,7 +2,7 @@
 FROM node:20.11.1
 
 # Set the working directory in the container
-WORKDIR /app/services
+WORKDIR /app
 
 # Copy package.json and package-lock.json to the container
 COPY package*.json ./
@@ -14,7 +14,7 @@ RUN npm install
 COPY . .
 
 # Expose port 3000 for the Node.js app
-EXPOSE 4200
+EXPOSE 3000
 
 # Define the command to run your Node.js app
-CMD ["node", "index.js"]
+CMD ["node", "server.js"]
